+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## 0.4.10 - 2017-03-21
+- show progress indicator for file searching
+- add a configuration key `maxFilesForSearch` to set the max files that allowed to search, default is 5120
+
 ## 0.4.9 - 2017-03-20
 - add `**/build/**` and `**/.vscode/**` into default exclude directories